+++ CHANGELOG.md
@@ -8,6 +8,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 ### Added
+- Players with [Player Locator Plus](https://modrinth.com/mod/player-locator-plus) installed now require a compass item in the inventory to display the bar. This can be disabled in the config.
+
+### Changed
+- Xaero's Minimap compat can now be disabled.
+
+## [1.0.1] - 04-20-2025
+
+### Added
 - Config option to disable effects moving overlays.
 
 ### Fixed
@@ -15,6 +23,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Dyed sacks now show overlays correctly.
 - Registry code loading too early on Forge.
 
-## 1.0.0
+## [1.0.0] - 04-17-2025
 
 Initial release.